+++ e/hu.e.compiler/src/hu/e/compiler/internal/model/TypeDefinitionResolver.java
@@ -38,7 +38,7 @@ public class TypeDefinitionResolver {
 		if (td instanceof ArrayTypeDef){
 			int baseSize = getSize(sm, ((ArrayTypeDef) td).getDef());
 			ILiteralSymbol length = (ILiteralSymbol)sm.resolve(((ArrayTypeDef)td).getSize());
-			return baseSize*length.getValue();
+			return baseSize*(int)length.getValue();
 		}
 		if (td instanceof StructTypeDef){
 			int size = 0;
