+++ e/hu.e.compiler/src/hu/e/compiler/internal/model/symbols/impl/OperationSymbol.java
@@ -154,7 +154,7 @@ public class OperationSymbol implements ILiteralSymbol, IVariableSymbol{
 	}
 	
 	@Override
-	public int getValue() throws ECompilerException {
+	public long getValue() throws ECompilerException {
 		if (isLiteral()){
 			if (this.a instanceof IReferenceSymbol) throw new ECompilerException(context, "Cannot execute operation on a link-time symbol!");
 			if (this.b instanceof IReferenceSymbol) throw new ECompilerException(context, "Cannot execute operation on a link-time symbol!");
