+++ e/hu.e.compiler/src/hu/e/compiler/internal/model/symbols/impl/VariableSymbol.java
@@ -85,7 +85,7 @@ public class VariableSymbol implements IVariableSymbol{
 		ArrayTypeDef atd = getArrayTypeDef(getType());
 		
 		if (atd == null) return null;
-		int length = ((ILiteralSymbol)sm.resolve(atd.getSize())).getValue();
+		int length = (int)((ILiteralSymbol)sm.resolve(atd.getSize())).getValue();
 
 		if (index < 0)
 			return null;
